requirements.txt
torch==1.1.0
torchvision==0.3.0
tqdm==4.42.1
matplotlib==3.1.3
numpy==1.16.0
pillow==6.2.0
pycocotools==2.0.0
tensorflow==1.15.2
opencv-python==4.2.0.32
Flask==1.1.1
Flask_SQLAlchemy==2.4.1
